extract course view holder (need refactoring)
