+++ app/src/main/java/org/stepic/droid/di/course_list/CourseListComponent.kt
@@ -4,6 +4,7 @@ import dagger.Subcomponent
 import org.stepic.droid.base.CoursesDatabaseFragmentBase
 import org.stepic.droid.ui.fragments.CourseListFragmentBase
 import org.stepic.droid.ui.fragments.CourseSearchFragment
+import org.stepic.droid.ui.fragments.HomeFragment
 
 @CourseListScope
 @Subcomponent(modules = arrayOf(CourseListModule::class))
@@ -20,4 +21,6 @@ interface CourseListComponent {
     fun inject(fragment: CourseListFragmentBase)
 
     fun inject(fragment: CourseSearchFragment)
+
+    fun inject(fragment: HomeFragment)
 }
