+++ app/src/main/java/org/stepic/droid/ui/fragments/HomeFragment.kt
@@ -1,22 +1,16 @@
 package org.stepic.droid.ui.fragments
 
 import android.os.Bundle
-import android.support.v7.widget.GridLayoutManager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import kotlinx.android.synthetic.main.fragment_home.*
 import org.stepic.droid.R
 import org.stepic.droid.base.FragmentBase
-import org.stepic.droid.ui.adapters.CourseSimpleAdapter
-import org.stepic.droid.ui.decorators.LeftSpacesDecoration
-import org.stepic.droid.ui.decorators.RightMarginForLastItems
-import org.stepic.droid.ui.decorators.VerticalSpacesForFirstRowDecoration
-import org.stepic.droid.ui.util.StartSnapHelper
-import org.stepic.droid.ui.util.initCenteredToolbar
 
-class HomeFragment : FragmentBase() {
+class HomeFragment : FragmentBase()//, ContinueCourseView, CoursesView, DroppingView {
+{
     companion object {
+
         fun newInstance(): HomeFragment {
             val args = Bundle()
             val fragment = HomeFragment()
@@ -26,27 +20,106 @@ class HomeFragment : FragmentBase() {
 
         //FIXME: 04.09.17 if adapter.count < ROW_COUNT -> recycler creates extra padding
         private const val ROW_COUNT = 2
-    }
 
+    }
+//
+//    @Inject
+//    lateinit var courseListPresenter: PersistentCourseListPresenter
+//
+//    @Inject
+//    lateinit var droppingPresenter: DroppingPresenter
+//
+//    @Inject
+//    lateinit var continueCoursePresenter: ContinueCoursePresenter
+//
+//    private var coursesAdapter: CoursesAdapter? = null
+//
+//    private val courses = ArrayList<Course>()
+//
+//    override fun injectComponent() {
+//        App
+//                .componentManager()
+//                .courseGeneralComponent()
+//                .courseListComponentBuilder()
+//                .build()
+//                .inject(this)
+//    }
+//
+//    override fun onReleaseComponent() {
+//        App
+//                .componentManager()
+//                .releaseCourseGeneralComponent()
+//    }
+//
     override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View?
             = inflater?.inflate(R.layout.fragment_home, container, false)
-
-    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        nullifyActivityBackground()
-        initCenteredToolbar(R.string.home_title)
-        initMyCoursesRecycler()
-    }
-
-    private fun initMyCoursesRecycler() {
-        myCoursesOnHome.layoutManager = GridLayoutManager(context, ROW_COUNT, GridLayoutManager.HORIZONTAL, false)
-        myCoursesOnHome.adapter = CourseSimpleAdapter(context)
-        val spacePx = resources.getDimensionPixelSize(R.dimen.course_list_between_items_padding)
-        myCoursesOnHome.addItemDecoration(VerticalSpacesForFirstRowDecoration(spacePx, ROW_COUNT))
-        myCoursesOnHome.addItemDecoration(LeftSpacesDecoration(spacePx))
-        myCoursesOnHome.addItemDecoration(RightMarginForLastItems(resources.getDimensionPixelSize(R.dimen.home_right_recycler_padding_without_extra), ROW_COUNT))
-        val snapHelper = StartSnapHelper()
-        snapHelper.attachToRecyclerView(myCoursesOnHome)
-    }
+//
+//    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
+//        super.onViewCreated(view, savedInstanceState)
+//        nullifyActivityBackground()
+//        initCenteredToolbar(R.string.home_title)
+//        initMyCoursesRecycler()
+//        continueCoursePresenter.attachView(this)
+//        courseListPresenter.attachView(this)
+//        droppingPresenter.attachView(this)
+//
+//        courseListPresenter.restoreState()
+//
+//        if (savedInstanceState == null) {
+//            courseListPresenter.downloadData(Table.enrolled, false)
+//        }
+//    }
+//
+//    override fun onDestroyView() {
+//        super.onDestroyView()
+//        continueCoursePresenter.detachView(this)
+//        courseListPresenter.detachView(this)
+//        droppingPresenter.detachView(this)
+//    }
+//
+//    private fun initMyCoursesRecycler() {
+//        myCoursesOnHome.layoutManager = GridLayoutManager(context, ROW_COUNT, GridLayoutManager.HORIZONTAL, false)
+//        myCoursesOnHome.adapter = CoursesAdapter(this, courses, Table.enrolled, continueCoursePresenter, droppingPresenter);
+//        val spacePx = resources.getDimensionPixelSize(R.dimen.course_list_between_items_padding)
+//        myCoursesOnHome.addItemDecoration(VerticalSpacesForFirstRowDecoration(spacePx, ROW_COUNT))
+//        myCoursesOnHome.addItemDecoration(LeftSpacesDecoration(spacePx))
+//        myCoursesOnHome.addItemDecoration(RightMarginForLastItems(resources.getDimensionPixelSize(R.dimen.home_right_recycler_padding_without_extra), ROW_COUNT))
+//        val snapHelper = StartSnapHelper()
+//        snapHelper.attachToRecyclerView(myCoursesOnHome)
+//    }
+//
+//    override fun onOpenStep(courseId: Long, section: Section, lessonId: Long, unitId: Long, stepPosition: Int) {
+////        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+//    }
+//
+//    override fun onAnyProblemWhileContinue(course: Course) {
+////        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+//    }
+//
+//    override fun onShowContinueCourseLoadingDialog() {
+////        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+//    }
+//
+//    override fun showLoading() {
+////        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+//    }
+//
+//    override fun showEmptyCourses() {
+////        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+//    }
+//
+//    override fun showConnectionProblem() {
+////        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+//    }
+//
+//    override fun showCourses(courses: MutableList<Course>) {
+//        this.courses.clear()
+//        this.courses.addAll(courses)
+//        coursesAdapter?.notifyDataSetChanged()
+//    }
+//
+//    override fun onUserHasNotPermissionsToDrop() {
+////        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+//    }
 
 }
