+++ app/src/main/res/layout/loading_view.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/loading_root"
+    android:id="@+id/loadingRoot"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:gravity="center"
